move SupportedVersions outside of v0
