+++ src/main/scala/nl/knaw/dans/bag/SupportedVersions.scala
@@ -1,4 +1,4 @@
-package nl.knaw.dans.bag.v0
+package nl.knaw.dans.bag
 
 import gov.loc.repository.bagit.domain.Version
